requirements.txt
Pillow~=8.4.0
pip~=21.1.2
wheel~=0.36.2
MarkupSafe~=2.0.1
Werkzeug~=2.0.2
Jinja2~=3.0.3
click~=8.0.3
itsdangerous~=2.0.1
numpy~=1.21.5
setuptools~=57.0.0
Flask~=2.0.2
